requirements.txt
# 核心科学计算库
numpy>=1.20.0
pandas>=1.2.0

# 优化求解器
gurobipy>=9.5.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 配置管理
pyyaml>=5.4.0

# 类型检查
typing-extensions>=3.10.0

# 日志
logging

# 开发和测试
pytest>=6.2.0
pytest-cov>=2.12.0
mypy>=0.910
black>=21.5b1

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.2

# 性能分析
memory-profiler>=0.58.0